api/requirements.txt
casparser[mupdf]==0.5.5
celery==5.4.0
Django==5.1.5
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-click>=2.3.0
django-environ==0.11.2
django-import-export==2.9.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
lxml
numpy
pandas
psycopg2
python-dateutil>=2.8.1
pytz>=2021.1
redis==5.2.1
requests==2.32.3
sentry-sdk==2.19.2
tablib[pandas]==3.7.0
xirr==0.1.8
